project upload callable as module or with comand line arguments
