--- helpers/project_upload.js
@@ -1,330 +1,0 @@
-// por_upload.js
-import { Synapse, RPC_URLS } from "@filoz/synapse-sdk";
-import { ethers } from "ethers";
-import fs from "fs";
-import path from "path";
-import { fileURLToPath } from "url";
-import archiver from "archiver";
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-
-class PorUploader {
-  constructor() {
-    this.synapse = null;
-  }
-
-  async initialize() {
-    const privateKey = process.env.PRIVATE_KEY;
-    if (!privateKey) {
-      console.error("❌ PRIVATE_KEY not found in environment variables");
-      console.log("Please create a .env file with your private key:");
-      console.log("PRIVATE_KEY=0x1234567890abcdef...");
-      process.exit(1);
-    }
-
-    console.log("Initializing Synapse SDK...");
-    const rpcURL = process.env.RPC_URL || RPC_URLS.calibration.http;
-    this.synapse = await Synapse.create({
-      privateKey: privateKey,
-      rpcURL: rpcURL,
-    });
-    console.log("✅ SDK initialized");
-  }
-
-  /**
-   * Ensures that the payment setup for the Synapse service is complete.
-   *
-   * This method performs the following steps:
-   * 1. Checks the current USDFC token balance.
-   * 2. Deposits additional USDFC tokens if the balance is below the required threshold.
-   * 3. Checks if the Warm Storage service is approved.
-   * 4. Approves the Warm Storage service if not already approved.
-   *
-   * If any step fails, logs an error message and suggests obtaining testnet USDFC tokens.
-   *
-   * @async
-   * @throws {Error} If payment setup fails at any step.
-   */
-  async ensurePaymentSetup() {
-    console.log("💰 Checking payment setup...");
-
-    try {
-      const balance = await this.synapse.payments.balance();
-      const balanceFormatted = ethers.formatUnits(balance, 18);
-      console.log(`Current USDFC balance: ${balanceFormatted}`);
-
-      const requiredBalance = ethers.parseUnits("2", 18);
-      if (balance < requiredBalance) {
-        console.log("💳 Depositing USDFC tokens...");
-        const depositAmount = ethers.parseUnits("5", 18);
-        await this.synapse.payments.deposit(depositAmount);
-        console.log("✅ Deposit complete");
-      }
-
-      const warmStorageAddress = this.synapse.getWarmStorageAddress();
-      const serviceStatus = await this.synapse.payments.serviceApproval(
-        warmStorageAddress
-      );
-
-      if (!serviceStatus.isApproved) {
-        console.log("🔐 Approving Warm Storage service...");
-        await this.synapse.payments.approveService(
-          warmStorageAddress,
-          ethers.parseUnits("10", 18),
-          ethers.parseUnits("1000", 18),
-          7776000n
-        );
-        console.log("✅ Service approval complete");
-      } else {
-        console.log("✅ Service already approved");
-      }
-    } catch (error) {
-      console.error("❌ Payment setup failed:", error.message);
-      console.log(
-        "💡 You may need testnet USDFC tokens. Check the Filecoin Calibration faucet."
-      );
-      throw error;
-    }
-  }
-
-  /**
-   * Creates a ZIP archive containing the specified files and returns the path to the created ZIP file.
-   *
-   * @async
-   * @param {string} por1Path - Path to the first file to include in the ZIP.
-   * @param {string} por2Path - Path to the second file to include in the ZIP.
-   * @param {string} scriptPath - Path to the script file to include in the ZIP.
-   * @returns {Promise<string>} Resolves with the path to the created ZIP file.
-   * @throws {Error} If any of the specified files do not exist or if an error occurs during ZIP creation.
-   */
-  async createZipFile(repoPath) {
-    const timestamp = Date.now();
-    const repoName = path.basename(repoPath);
-    const files = fs.readdirSync(repoPath);
-    const zipPath = path.join(__dirname, `${repoName}_${timestamp}.zip`);
-
-    console.log("📦 Creating ZIP bundle...");
-
-    return new Promise((resolve, reject) => {
-      const output = fs.createWriteStream(zipPath);
-      const archive = archiver("zip", { zlib: { level: 9 } });
-
-      output.on("close", () => {
-        console.log(`✅ ZIP created: ${archive.pointer()} bytes`);
-        resolve(zipPath);
-      });
-
-      archive.on("error", (err) => reject(err));
-      archive.pipe(output);
-
-      files.forEach((file) => {
-        const filePath = path.join(repoPath, file);
-        const stat = fs.statSync(filePath);
-
-        if (stat.isFile()) {
-          archive.file(filePath, { name: file });
-          console.log(`  📄 Added: ${file}`);
-        } else if (stat.isDirectory()) {
-          archive.directory(filePath, file);
-          console.log(`  📁 Added directory: ${file}`);
-        }
-      });
-
-      archive.finalize();
-    });
-  }
-  /**
-   * Creates a ZIP bundle from the provided file paths, uploads it to Filecoin via Synapse storage,
-   * and returns the upload result including the bundle CID, size, and compression ratio.
-   *
-   * @async
-   * @param {string} por1Path - Path to the first POR file.
-   * @param {string} por2Path - Path to the second POR file.
-   * @param {string} scriptPath - Path to the script file to include in the bundle.
-   * @returns {Promise<Object>} Resolves with an object containing:
-   *   - {string} bundleCid - The CID of the uploaded bundle.
-   *   - {number} size - The size of the uploaded bundle in bytes.
-   *   - {string} compressionRatio - The compression ratio as a percentage string.
-   * @throws Will throw an error if ZIP creation or upload fails.
-   */
-  async uploadZipBundle(repoPath) {
-    try {
-      // Create ZIP bundle
-      const zipPath = await this.createZipFile(repoPath);
-
-      // Upload the ZIP
-      console.log("📤 Uploading ZIP bundle to Filecoin...");
-      const zipData = fs.readFileSync(zipPath);
-      const zipSize = zipData.length;
-
-      console.log(`📦 Uploading bundle (${zipSize} bytes)...`);
-      const uploadResult = await this.synapse.storage.upload(zipData);
-
-      // Clean up temporary ZIP
-      fs.unlinkSync(zipPath);
-
-      console.log("✅ ZIP bundle uploaded successfully!");
-
-      return {
-        bundleCid: uploadResult.pieceCid,
-        size: uploadResult.size,
-        compressionRatio: ((1 - uploadResult.size / zipSize) * 100).toFixed(1),
-      };
-    } catch (error) {
-      console.error("❌ Failed to create or upload ZIP bundle:", error.message);
-      throw error;
-    }
-  }
-
-  async dlHuggingFaceRepo(repoUrl, destDir = "./models") {
-    const { exec } = await import("child_process");
-    return new Promise((resolve, reject) => {
-      console.log(`\n🔽 Cloning Hugging Face repo: ${repoUrl}`);
-      const repoName = repoUrl.split("/").pop();
-      exec(
-        `git clone ${repoUrl} ${destDir}/${repoName}`,
-        (error, stdout, stderr) => {
-          if (error) {
-            console.error(`❌ Error cloning repo: ${error.message}`);
-            return reject(error);
-          }
-          console.log(`✅ Cloned repo successfully: ${stdout}`);
-          resolve();
-        }
-      );
-    });
-  }
-}
-
-async function createRepoSnapshot(repoPath) {
-  // Get all files in the models directory
-  const files = fs
-    .readdirSync(repoPath)
-    .map((file) => path.basename(file))
-    .join("; ");
-
-  console.log(`📂 Found files: ${files}`);
-}
-
-/**
- * Uploads a ZIP bundle containing the specified PoR files and script, and logs upload details.
- * This function initializes the uploader, ensures payment setup, creates the ZIP bundle,
- * and returns the upload result including the bundle CID, size, and compression ratio.
- *
- * @async
- * @function runUpload
- * @param {string} por1Path - Path to the first PoR file.
- * @param {string} por2Path - Path to the second PoR file.
- * @param {string} scriptPath - Path to the script file to include in the bundle.
- * @returns {Promise<void>} Resolves when the upload process is complete.
- */
-async function runUpload(repoUrl) {
-  const uploader = new PorUploader();
-
-  try {
-    await uploader.initialize();
-    await uploader.ensurePaymentSetup();
-
-    console.log("\n🔽 Downloading Hugging Face repo...");
-    await uploader.dlHuggingFaceRepo(repoUrl);
-    const repoName = repoUrl.split("/").pop();
-    const repoPath = path.join("./models", repoName);
-
-    const startTime = Date.now();
-
-    console.log("📤 Uploading ZIP bundle to Filecoin...");
-    const result = await uploader.uploadZipBundle(repoPath);
-
-    const endTime = Date.now();
-    const uploadDuration = (endTime - startTime) / 1000;
-
-    console.log("\n🎉 Upload Complete!");
-    console.log("===================");
-    console.log(`📦 Bundle CID: ${result.bundleCid}`);
-    console.log(`📊 Bundle Size: ${result.size} bytes`);
-    console.log(`⏱️  Upload Time: ${uploadDuration.toFixed(2)} seconds`);
-
-    console.log("\n💾 Single CID for entire PoR bundle:");
-    console.log(`   ${result.bundleCid}`);
-    console.log(
-      "\n📥 To retrieve: Download this CID to get a ZIP with all files!"
-    );
-  } catch (error) {
-    console.error("\n❌ Upload failed:", error.message);
-    process.exit(1);
-  }
-}
-
-// CLI handling
-async function main() {
-  console.log("Cairn PoR ZIP Bundle Uploader");
-  console.log("==============================");
-
-  const args = process.argv.slice(2);
-
-  if (args.length === 0) {
-    console.log(`
-Usage Options:
-
-1. Command line arguments:
-   node project_upload.js <hugging face repo url>
-
-2. Interactive mode (no arguments):
-   node project_upload.js
-
-Examples:
-   node project_upload.js https://huggingface.co/username/repo_name
-   node project_upload.js  # Interactive mode
-
-Features:
-   - Creates a single ZIP containing all files
-   - One upload = One CID for the entire repo
-   - Easy to download and extract later
-
-Environment Setup:
-   Create a .env file with: PRIVATE_KEY=0x1234567890abcdef...
-    `);
-
-    const readline = await import("readline");
-    const rl = readline.createInterface({
-      input: process.stdin,
-      output: process.stdout,
-    });
-
-    const question = (prompt) =>
-      new Promise((resolve) => rl.question(prompt, resolve));
-
-    try {
-      console.log("\n📝 Interactive Mode - Enter file paths:");
-      const hugging_face_repo_url = await question("Hugging Face repo URL: ");
-
-      rl.close();
-      await runUpload(hugging_face_repo_url.trim());
-    } catch (error) {
-      rl.close();
-      throw error;
-    }
-    return;
-  }
-
-  if (args.length !== 1) {
-    console.error("❌ Please provide exactly 1 Hugging Face repo URL");
-    process.exit(1);
-  }
-
-  await runUpload(args[0]);
-}
-
-// Load environment variables
-import dotenv from "dotenv";
-dotenv.config();
-
-// Run if called directly
-// main().catch(console.error);
-
-const uploader = new PorUploader();
-await uploader.initialize();
-await createRepoSnapshot(
-  "C:\\Users\\lukal\\Dev\\cairn\\helpers\\models\\detr-resnet-50"
-);
